(NOBIDS) reorganize code
